packages/streamlib-extras/pyproject.toml
[project]
name = "streamlib-extras"
version = "0.1.0"
description = "Reference handler implementations for streamlib"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "streamlib contributors"}
]
keywords = ["streaming", "video", "handlers", "gpu"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

# Core dependency: streamlib SDK
dependencies = [
    "streamlib>=0.2.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
# Basic handlers (CPU-based)
basic = [
    "opencv-python>=4.8.0",
]

# GPU handlers
gpu = [
    "torch>=2.0.0",
]

# GPU display
display = [
    "moderngl>=5.8.0",
    "glfw>=2.6.0",
]

# Text overlays
text = [
    "Pillow>=10.0.0",
]

# Metal backend (macOS only)
metal = [
    "pyobjc-framework-AVFoundation>=9.0",
    "pyobjc-framework-CoreMedia>=9.0",
    "pyobjc-framework-Metal>=9.0",
]

# All features
all = [
    "opencv-python>=4.8.0",
    "torch>=2.0.0",
    "moderngl>=5.8.0",
    "glfw>=2.6.0",
    "Pillow>=10.0.0",
]

[project.urls]
Homepage = "https://github.com/tatolab/gst-mcp-tools"
Repository = "https://github.com/tatolab/gst-mcp-tools"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/streamlib_extras"]
